[FEAT] 관심등록이 되어있지 않은 userNovel만 삭제하도록 처리

+++ src/main/java/org/websoso/WSSServer/service/UserNovelService.java
@@ -93,7 +93,11 @@ public class UserNovelService {
         List<NovelKeyword> novelKeywords = novelKeywordRepository.findAllByNovelAndUserId(novel, user.getUserId());
         novelKeywordRepository.deleteAll(novelKeywords);
 
-        userNovelRepository.delete(userNovel);
+        if (!userNovel.getIsInterest()) {
+            userNovelRepository.delete(userNovel);
+        }
+
+        userNovel.deleteEvaluation();
     }
 
     public UserNovel createUserNovelByInterest(User user, Novel novel) {
